germplasm import: move list fields into germplasm object
