+++ src/main/java/org/breedinginsight/brapps/importer/model/base/BrAPIList.java
@@ -16,21 +16,4 @@ public class BrAPIList implements BrAPIObject {
     private String listDescription;
     private BrAPIListTypes listType;
     private List<ExternalReference> externalReferences;
-
-    public BrAPIListNewRequest constructBrAPIList(Program program, String referenceSource) {
-        BrAPIListNewRequest brapiList = new BrAPIListNewRequest();
-        brapiList.setListName(constructGermplasmListName(listName, program));
-        brapiList.setListDescription(this.listDescription);
-        brapiList.listType(listType);
-        // Set external reference
-        BrAPIExternalReference reference = new BrAPIExternalReference();
-        reference.setReferenceSource(String.format("%s/programs", referenceSource));
-        reference.setReferenceID(program.getId().toString());
-        brapiList.setExternalReferences(List.of(reference));
-        return brapiList;
-    }
-
-    public static String constructGermplasmListName(String listName, Program program) {
-        return String.format("%s [%s-germplasm]", listName, program.getName());
-    }
 }
